+++ src/inferenceresponse.hpp
@@ -37,7 +37,7 @@ public:
     // this constructor can be removed with prediction tests overhaul
     InferenceResponse();
     InferenceResponse(const std::string& servableName, model_version_t servableVersion);
-    Status addOutput(const std::string& name, OVMS_DataType datatype, const size_t* shape, size_t dimCount);
+    Status addOutput(const std::string& name, OVMS_DataType datatype, const int64_t* shape, size_t dimCount);
     Status getOutput(uint32_t id, const std::string** name, const InferenceTensor** tensor) const;
     Status getOutput(uint32_t id, const std::string** name, InferenceTensor** tensor);
